Move addActivity() to commons.ts
